weather_station/requirements.txt
smbus2
RPi.bme280
pyserial
pynmea2
